+++ lib/widgets/about/translators.dart
@@ -42,10 +42,10 @@ class AboutTranslators extends StatelessWidget {
     Contributor('vesp', 'vesp@post.cz'),
     Contributor('Dan', 'denqwerta@gmail.com'),
     Contributor('Tijolinho', 'pedrohenrique29.alfenas@gmail.com'),
+    Contributor('Piotr K', '1337.kelt@gmail.com'),
+    Contributor('rehork', 'cooky@e.email'),
     // Contributor('SAMIRAH AIL', 'samiratalzahrani@gmail.com'), // Arabic
     // Contributor('Salih Ail', 'rrrfff444@gmail.com'), // Arabic
-    // Contributor('Piotr K', '1337.kelt@gmail.com'), // Polish
-    // Contributor('rehork', 'cooky@e.email'), // Polish
     // Contributor('امیر جهانگرد', 'ijahangard.a@gmail.com'), // Persian
     // Contributor('slasb37', 'p84haghi@gmail.com'), // Persian
     // Contributor('tryvseu', 'tryvseu@tuta.io'), // Nynorsk