+++ lib/widgets/settings/language/locales.dart
@@ -16,6 +16,7 @@ class SupportedLocales {
     'lt': 'Lietuvių',
     'nb': 'Norsk (Bokmål)',
     'nl': 'Nederlands',
+    'pl': 'Polski',
     'pt': 'Português (Brasil)',
     'ro': 'Română',
     'ru': 'Русский',